FIX: ilo_progress_summary will not return an error
student_ilo_progress_stats.values previously calculated as

{10=>4.8, 11=>4.8, 12=>3.2}
{10=>0.8, 11=>0.8, 12=>1.6}
{10=>0.8, 11=>0.8}

There is no ILO id 12 for third member

Causes ilo_progress_summary to fail on the sort, as the map of the data
of ILO 12 would be [3.2, 1.6, nil]

Defaults to 0 now instead of nil. Not the best fix but at least we can get
some sort of data back.