FIX: Push lazy project tasks to project.tasks if created
When a task for task_for_task_definition is lazily created, it was
created but wasn't adding the task to the project.task array.

This was causing tasks generated using the task_for_task_definition
to create tasks but not have the project "know" about those tasks.
Whilst the task knew about the project, the project's task array
would always be empty (especially when running rake db:simulate_signoff)