FIX: Add missing RAILS_ENV for docker api
This is needed since when the rake db:populate command is run,
it will try and run for BOTH the test and development databases.
This is because since db:populate will run db:create, which by
default will try and create the test database also. Refer to
See http://jacopretorius.net/2014/02/all-rails-db-rake-tasks-
and-what-they-do.html

The test database still points to localhost, and since there is
no way to have multiple DATABASE_URL's for test and development,
this can only be overriden by having a static RAILS_ENV value.