fix: truncate long lines in PDF conversion (#439)
* chore: add shellwords gem for proper shell path escape

* fix: process code files to remove long lines

PDF processing will fail with "! Dimension too large." if a line of text
is way too long. Implement a simple processing helper method to call
fold (from coreutils) to fold long lines for all code files.

* refactor: write line-wrapped code files into temp files

We shouldn't modify student submissions, instead we use the
temp file when rendering them to PDF. Cleanup will be performed
after rendering is complete.

* enhance: render line-wrapped code files instead

Run fold on provided files and compare the output with diff. If the file
doesn't contain any lines that are over the configured threshold, it
will be identical to the original. In this case we replace the temp file
with a symlink for easy identification in the template.

* fix: silence diff

Redirect stdout to /dev/null since we don't need the diff output

* fix: change default column width limit to 160 characters

* enhance: add a notice in the pdf template if file is modified

* enhance: truncate super long lines in code files rendered in pdf

Set a hard limit of 1000 characters, and truncate everything in the same
line after the limit is reached. Otherwise we could get PDF files with
hundreds of pages which is completely unreadable.

* refactor: update temp file names

* quality: add unit test for code submissions with long lines

* quality: ensure long line notice is not included when not applicable

* fix: update comment to reflect temp filename changes

* refactor: avoid making temp files since we are working with copies

* chore: update jupynotex.py to latest version

* fix: truncate unreasonably long lines (1000 char) in source cells

* fix: allow generation of notebooks with one cell (upstream bug)

* fix: force line wraps for source cells

* quality: test pdf conversion on ipynb with long lines

* fix: replace deprecated option and silence useless errors for inkscape

* enahnce: replace inkscape with librsvg

https://gitlab.gnome.org/GNOME/librsvg

---------

Co-authored-by: Andrew Cain <macite@gmail.com>